+++ engine/flink/components-api/src/main/scala/pl/touk/nussknacker/engine/flink/api/process/FlinkSource.scala
@@ -6,6 +6,7 @@ import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
 import pl.touk.nussknacker.engine.api.Context
 import pl.touk.nussknacker.engine.api.process.{BasicContextInitializer, ContextInitializer, Source, SourceTestSupport}
 import pl.touk.nussknacker.engine.api.typed.typing.Unknown
+import pl.touk.nussknacker.engine.flink.api.compat.ExplicitUidInOperatorsSupport
 import pl.touk.nussknacker.engine.flink.api.timestampwatermark.TimestampWatermarkHandler
 
 /**
@@ -37,7 +38,7 @@ trait FlinkSourceTestSupport[Raw] extends SourceTestSupport[Raw] with ExplicitTy
 }
 
 /**
- * Standard source providing default logic for transforming a `DataStreamSource[Raw]` into `DataStream[Context]` using
+ * Source providing standard logic for transforming a `DataStreamSource[Raw]` into `DataStream[Context]` using
  * customizable [[ContextInitializer]] and [[TimestampWatermarkHandler]].
  *
  * @tparam Raw - type of raw event that is generated by flink source function.
@@ -45,7 +46,8 @@ trait FlinkSourceTestSupport[Raw] extends SourceTestSupport[Raw] with ExplicitTy
 trait StandardFlinkSource[Raw]
     extends FlinkSource
     with CustomizableContextInitializerSource[Raw]
-    with CustomizableTimestampWatermarkHandlerSource[Raw] {
+    with CustomizableTimestampWatermarkHandlerSource[Raw]
+    with ExplicitUidInOperatorsSupport {
 
   def sourceStream(
       env: StreamExecutionEnvironment,
@@ -56,12 +58,28 @@ trait StandardFlinkSource[Raw]
       env: StreamExecutionEnvironment,
       flinkNodeContext: FlinkCustomNodeContext
   ): DataStream[Context] = {
-    FlinkStandardSourceUtils.createContextStream(
-      source = sourceStream(env, flinkNodeContext),
-      flinkNodeContext = flinkNodeContext,
-      timestampAssigner = timestampAssigner,
-      contextInitializer = contextInitializer
+    // 1. set UID and override source name
+    val rawSourceWithUid = setUidToNodeIdIfNeed[Raw](
+      flinkNodeContext,
+      sourceStream(env, flinkNodeContext)
+        .name(flinkNodeContext.nodeId)
     )
+
+    // 2. assign timestamp and watermark policy
+    val rawSourceWithUidAndTimestamp = timestampAssigner
+      .map(_.assignTimestampAndWatermarks(rawSourceWithUid))
+      .getOrElse(rawSourceWithUid)
+
+    // 3. initialize Context and spool Context to the stream
+    rawSourceWithUidAndTimestamp
+      .map(
+        new FlinkContextInitializingFunction(
+          contextInitializer,
+          flinkNodeContext.nodeId,
+          flinkNodeContext.convertToEngineRuntimeContext
+        ),
+        flinkNodeContext.contextTypeInfo
+      )
   }
 
 }
